chore: enhance destroy-cluster action to validate kubeconfig size before deletion

+++ src/qa-test-eks.sh
@@ -61,8 +61,9 @@ if [ "$ACTION" = "destroy-cluster" ]; then
 	aws ssm get-parameter --name /eks/qa/kubeconfig-admin --with-decryption --region eu-west-1 --query 'Parameter.Value' --output text >$PWD/qa.yaml || true
 
 	export KUBECONFIG=$PWD/qa.yaml
+	export KUBECONFIG_SIZE=$(du -k $PWD/qa.yaml | cut -f 1)
 
-	if [[ -f $KUBECONFIG ]]; then
+	if [[ -f $KUBECONFIG && $KUBECONFIG_SIZE -gt 0 ]]; then
 		kubectl delete APIServices v1beta1.metrics.k8s.io
 
 		NAMESPACES=$(kubectl get namespaces --no-headers -o custom-columns=NAME:.metadata.name | awk '{print $1}' | tail -n +2)
